First run at Magento site2mysql
The script makes a call for each product ID (they start at 139)
and runs into the hard coded end of products. This is because
compiling the ENTIRE magento DB causes the server to run out
of allowed memory for the API calls. Not sure if there's a work
around or if we're stuck with this one-by-one concept. The first
few products came down in the sample run no problem, and are built
into a hash in memory before I spit them out into a file for testing
purposes. We need to examine the keys of the hash and decide what to
do with them. The CSV file that Magento spits out has 103 columns,
this process only has 63... I haven't put together what the missing
40 columns are and if there are needed.